+++ R/summary.R
@@ -925,8 +925,10 @@ print.summary_emm = function(x, ..., digits=NULL, quote=FALSE, right=TRUE) {
         if (is.numeric(x[[i]]))  
             x[[i]] = zapsmall(x[[i]])
     if (!is.null(x$df)) x$df = round(x$df, 2)
-    if (!is.null(x$t.ratio)) x$t.ratio = round(x$t.ratio, 3)
-    if (!is.null(x$z.ratio)) x$z.ratio = round(x$z.ratio, 3)
+    if (!is.null(x$t.ratio)) 
+        x$t.ratio = format(round(x$t.ratio, 3), nsmall = 3, sci = FALSE)
+    if (!is.null(x$z.ratio)) 
+        x$z.ratio = format(round(x$z.ratio, 3), nsmall = 3, sci = FALSE)
     if (!is.null(x$p.value)) {
         fp = x$p.value = format(round(x$p.value,4), nsmall=4, sci=FALSE)
         x$p.value[fp=="0.0000"] = "<.0001"
@@ -941,7 +943,7 @@ print.summary_emm = function(x, ..., digits=NULL, quote=FALSE, right=TRUE) {
             tmp = x[, c("lower.HPD", estn, "upper.HPD"), drop = FALSE]
         else tmp = NULL
         if (!is.null(tmp))
-            digits = max(apply(tmp, 1, .opt.dig), na.rm = TRUE)
+            digits = max(apply(tmp, 1, .opt.dig))
     }
     if (any(is.na(est))) {
         x[[estn]] = format(est, digits=digits)
@@ -985,7 +987,7 @@ print.summary_emm = function(x, ..., digits=NULL, quote=FALSE, right=TRUE) {
 # (but always at least 3)
 .opt.dig = function(x) {
     z = range(x) / max(abs(x))
-    max(round(1.51 - log(diff(z), 10)), 3)  # approx 1 - log(diff(z/3))
+    max(round(1.51 - log(diff(z), 10)), 3, na.rm = TRUE)  # approx 1 - log(diff(z/3))
 }
 
 # Utility -- When misc$display present, reconcile which elements to use.