+++ NEWS.md
@@ -1,4 +1,6 @@
-emmeans 1.3.0.09999x
+## NEWS for the emmeans package
+
+emmeans 1.3.1
 -------------
 
   * `"mvt"` adjustment ignored `by` grouping
@@ -7,7 +9,9 @@ emmeans 1.3.0.09999x
   * Changed `aovlist` support so it re-fits the model when non-sum-to-zero
     contrasts were used
   * `print.summary_emm()` now cleans up numeric columns with `zapsmall()`
-  * More robust handling of `nesting` in `ref_grid()` and `update()`
+  * More robust handling of `nesting` in `ref_grid()` and `update()`,
+    and addition of `covnest` argument for whether to include covariates
+    when auto-detecting nesting
   * Revision of some vignettes
   * Fixed bug in `hpd.summary()` and handoff to it from `summary()`
   * Fixed bug where `ref_grid()` ignored `mult.levs`